+++ packages/components/layout/sidebar/Sidebar.tsx
@@ -85,6 +85,7 @@ type SubItem = {
   value: string;
   label: string;
   slug?: string;
+  badge?: { label: string; color: ChipColors };
   icon?: any;
   onMouseDown?: (e: React.MouseEvent) => void;
   onClick?: (e: React.MouseEvent) => void;
@@ -203,6 +204,9 @@ const SidebarItem: React.FC<{
               )}
             >
               {item.label}
+              {item.badge && (
+                <Chip label={item.badge.label} color="hyper" size="small" />
+              )}
             </span>
           </div>
         </AccordionTrigger>
@@ -239,6 +243,13 @@ const SidebarItem: React.FC<{
                 >
                   {subitem.icon && subitem.icon}
                   {subitem.label}
+                  {subitem.badge && (
+                    <Chip
+                      label={subitem.badge.label}
+                      color="hyper"
+                      size="small"
+                    />
+                  )}
                 </LinkComponent>
               ))}
             </div>